const replaced by readonly properties
