+++ src/Selenium.WebDriver.Extensions/Loaders/JQueryLoader.cs
@@ -3,7 +3,7 @@
     using System;
     using System.Diagnostics.CodeAnalysis;
     using System.Linq;
-    using static OpenQA.Selenium.Loaders.JavaScriptSnippets;
+    using static OpenQA.Selenium.JavaScriptSnippets;
 
     /// <summary>
     /// The jQuery loader.
